// src/core/search/types.ts
export interface SearchResult {
  id: string
  title: string
  subtitle: string
  module: string
  moduleColor: string
  actionUrl: string
  icon: string
}

export interface SearchIndex {
  id: string
  title: string
  subtitle: string
  module: string
  moduleColor: string
  actionUrl: string
  icon: string
  keywords: string
}